requirements.txt
datasets
pandas
numpy
torch
transformers
tqdm
huggingface_hub
requests
python-dotenv
google-genai
mistralai
tabulate
plotly
